--- a/hw05_parallel_execution/go.mod
+++ b/hw05_parallel_execution/go.mod
@@ -1,13 +1,13 @@
-module github.com/fixme_my_friend/hw05_parallel_execution
+module github.com/giiuser/hw-otus/hw05_parallel_execution
 
 go 1.16
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/stretchr/testify v1.7.0
-	go.uber.org/goleak v1.1.10
+	go.uber.org/goleak v1.1.12
+	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9 // indirect
 	golang.org/x/lint v0.0.0-20201208152925-83fdc39ff7b5 // indirect
-	golang.org/x/tools v0.1.0 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )