--- a/hw07_file_copying/go.mod
+++ b/hw07_file_copying/go.mod
@@ -1,3 +1,8 @@
 module github.com/giiuser/hw-otus/hw07_file_copying
 
 go 1.16
+
+require (
+	github.com/cheggaaa/pb/v3 v3.0.8 // indirect
+	github.com/stretchr/testify v1.7.0 // indirect
+)