--- a/hw07_file_copying/main.go
+++ b/hw07_file_copying/main.go
@@ -18,5 +18,9 @@ func init() {
 
 func main() {
 	flag.Parse()
-	// Place your code here.
+
+	err := Copy(from, to, offset, limit)
+	if err != nil {
+		return
+	}
 }