--- a/hw08_envdir_tool/go.mod
+++ b/hw08_envdir_tool/go.mod
@@ -1,3 +1,5 @@
-module github.com/fixme_my_friend/hw08_envdir_tool
+module github.com/giiuser/hw-otus/hw08_envdir_tool
 
 go 1.16
+
+require github.com/stretchr/testify v1.7.0 // indirect