--- a/hw08_envdir_tool/main.go
+++ b/hw08_envdir_tool/main.go
@@ -1,5 +1,19 @@
 package main
 
+import (
+	"log"
+	"os"
+)
+
 func main() {
-	// Place your code here.
+	args := os.Args
+
+	envs, err := ReadDir(args[1])
+	if err != nil {
+		log.Println(err)
+		os.Exit(1)
+	}
+
+	n := RunCmd(args[2:], envs)
+	os.Exit(n)
 }