--- a/hw09_struct_validator/go.mod
+++ b/hw09_struct_validator/go.mod
@@ -1,3 +1,5 @@
-module github.com/fixme_my_friend/hw09_struct_validator
+module github.com/giiuser/hw-otus/hw09_struct_validator
 
 go 1.16
+
+require github.com/stretchr/testify v1.7.0 // indirect