--- a/hw10_program_optimization/stats_optimization_test.go
+++ b/hw10_program_optimization/stats_optimization_test.go
@@ -1,3 +1,5 @@
+// +build bench
+
 package hw10programoptimization
 
 import (