… hw05_parallel_execution/go.sum
… hw07_file_copying/go.sum
… hw08_envdir_tool/go.sum
… hw09_struct_validator/go.sum
… hw10_program_optimization/go.sum